feat(frontend): migrate run-detail, checks, diff pages to TanStack Query #78
Phase 4: Complex pages migration
- Run detail page: 5 prefetchQuery (run, artifacts, viewer-sources, project, diff)
  + New RunDetailContent client component with useSuspenseQuery
  + diff uses useQuery (non-suspense) for graceful 404 handling
- Checks/findings page: 2 prefetchQuery (findings, project)
  + Validation (checkKind, severity) remains in Server Component
  + New FindingsContent client component
- Diff page: 2 prefetchQuery (diff, project)
  + New DiffContent client component with all section sub-components
- All pages use HydrationBoundary + Suspense pattern
- Build passes with no TypeScript errors
